+++ authentication-service/src/main/kotlin/com/saveourtool/save/authservice/utils/AuthenticationUtils.kt
@@ -5,7 +5,6 @@
 package com.saveourtool.save.authservice.utils
 
 import com.saveourtool.save.domain.Role
-import com.saveourtool.save.utils.AuthenticationDetails
 import org.springframework.security.access.hierarchicalroles.RoleHierarchy
 import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl
 import org.springframework.security.access.hierarchicalroles.RoleHierarchyUtils
