+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/controllers/ProjectController.kt
@@ -12,7 +12,7 @@ import com.saveourtool.save.domain.Role
 import com.saveourtool.save.entities.*
 import com.saveourtool.save.filters.ProjectFilters
 import com.saveourtool.save.permission.Permission
-import com.saveourtool.save.utils.AuthenticationDetails
+import com.saveourtool.save.authservice.utils.AuthenticationDetails
 import com.saveourtool.save.v1
 
 import io.swagger.v3.oas.annotations.Operation
