+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/controllers/internal/UsersController.kt
@@ -4,8 +4,8 @@ import com.saveourtool.save.backend.StringResponse
 import com.saveourtool.save.backend.repository.OriginalLoginRepository
 import com.saveourtool.save.backend.service.UserDetailsService
 import com.saveourtool.save.entities.User
-import com.saveourtool.save.utils.IdentitySourceAwareUserDetails
-import com.saveourtool.save.utils.IdentitySourceAwareUserDetailsMixin
+import com.saveourtool.save.authservice.utils.IdentitySourceAwareUserDetails
+import com.saveourtool.save.authservice.utils.IdentitySourceAwareUserDetailsMixin
 import com.saveourtool.save.utils.extractUserNameAndSource
 
 import com.fasterxml.jackson.databind.ObjectMapper
