+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/service/LnkUserProjectService.kt
@@ -7,7 +7,7 @@ import com.saveourtool.save.entities.LnkUserProject
 import com.saveourtool.save.entities.Project
 import com.saveourtool.save.entities.ProjectStatus
 import com.saveourtool.save.entities.User
-import com.saveourtool.save.utils.AuthenticationDetails
+import com.saveourtool.save.authservice.utils.AuthenticationDetails
 import com.saveourtool.save.utils.getHighestRole
 
 import org.springframework.data.domain.PageRequest
