+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/service/UserDetailsService.kt
@@ -6,7 +6,7 @@ import com.saveourtool.save.domain.Role
 import com.saveourtool.save.domain.UserSaveStatus
 import com.saveourtool.save.entities.OriginalLogin
 import com.saveourtool.save.entities.User
-import com.saveourtool.save.utils.getIdentitySourceAwareUserDetails
+import com.saveourtool.save.authservice.utils.getIdentitySourceAwareUserDetails
 import com.saveourtool.save.utils.orNotFound
 import org.slf4j.LoggerFactory
 import org.springframework.security.core.Authentication
