+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/utils/ConvertingAuthenticationManager.kt
@@ -1,8 +1,8 @@
 package com.saveourtool.save.backend.utils
 
 import com.saveourtool.save.backend.service.UserDetailsService
-import com.saveourtool.save.utils.AuthenticationDetails
-import com.saveourtool.save.utils.IdentitySourceAwareUserDetails
+import com.saveourtool.save.authservice.utils.AuthenticationDetails
+import com.saveourtool.save.authservice.utils.IdentitySourceAwareUserDetails
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.security.authentication.BadCredentialsException
 import org.springframework.security.authentication.ReactiveAuthenticationManager
