+++ save-cloud-common/build.gradle.kts
@@ -43,7 +43,6 @@ kotlin {
         val jvmMain by getting {
             dependencies {
                 implementation(project.dependencies.platform(libs.spring.boot.dependencies))
-                implementation(libs.spring.security.core)
                 implementation(libs.spring.web)
                 implementation(libs.spring.boot)
                 implementation(libs.spring.data.jpa)
